Moved conversationPage folder into components
